+++ ts_netmon/src/family.rs
@@ -1,4 +1,7 @@
-use core::fmt::{Debug, Formatter};
+use core::{
+    fmt::{Debug, Formatter},
+    net::{IpAddr, Ipv4Addr, Ipv6Addr},
+};
 
 /// Specification of IPv4 or IPv6 or both.
 #[derive(Copy, Clone, PartialEq, Eq)]
@@ -10,7 +13,7 @@ pub enum FamilyOrBoth {
 }
 
 impl Debug for FamilyOrBoth {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         match self {
             FamilyOrBoth::Both => write!(f, "Both"),
             FamilyOrBoth::Single(family) => family.fmt(f),
@@ -41,5 +44,25 @@ impl TryFrom<FamilyOrBoth> for Family {
             FamilyOrBoth::Both => Err(()),
             FamilyOrBoth::Single(family) => Ok(family),
         }
+    }
+}
+
+impl From<IpAddr> for Family {
+    fn from(addr: IpAddr) -> Self {
+        match addr {
+            IpAddr::V4(_) => Family::Ipv4,
+            IpAddr::V6(_) => Family::Ipv6,
+        }
+    }
+}
+impl From<Ipv4Addr> for Family {
+    fn from(_: Ipv4Addr) -> Self {
+        Family::Ipv4
+    }
+}
+
+impl From<Ipv6Addr> for Family {
+    fn from(_: Ipv6Addr) -> Self {
+        Family::Ipv6
     }
 }